chore: migrate pubsub module into the gcloud module
