--- docs/examples/pubsub.md
@@ -1,9 +1,0 @@
-# Pubsub
-
-<!--codeinclude-->
-[Creating a Pubsub container](../../examples/pubsub/pubsub.go)
-<!--/codeinclude-->
-
-<!--codeinclude-->
-[Test for a Pubsub container](../../examples/pubsub/pubsub_test.go)
-<!--/codeinclude-->
