--- examples/pubsub/Makefile
@@ -1,5 +1,0 @@
-include ../../commons-test.mk
-
-.PHONY: test
-test:
-	$(MAKE) test-pubsub
